camara cambio de pos logica a pos fisica, anadido UICamera a scene

+++ Carom/src/SceneSystem/Camera.h
@@ -4,13 +4,13 @@ class b2Vec2;
 
 class Camera{
 public:
-    //su posicion es publica porque me parece mas comoda para animarla asi sin usarla getters ni setters
-    int x =0;
-    int y =0;
+    //Posicion fisica. su posicion es publica porque me parece mas comoda para animarla asi sin usarla getters ni setters
+    float x = 0;
+    float y = 0;
 
-    Camera(int nX, int nY);
+    Camera(float nX, float nY);
     //setea la posicion de la camara con una linea
-    void setPos(int newX, int newY);
+    void setPos(float newX, float newY);
     //convierte la posicion fisica en posicion en la pantalla DEPENDIENTE de la posicion de la camara
     std::pair<int,int> getRenderPos(b2Vec2 physicalPos);
 };
